Rename file names in backend and shared folders
