+++ backend/src/routes/chat/chat.ts
@@ -1,6 +1,6 @@
 import { FastifyInstance } from "fastify";
 import { Static } from "@sinclair/typebox";
-import { getApiKey } from "../../controllers/apiKey";
+import { getApiKey } from "../../controllers/api-key";
 import {
   ChatCompletionResponse,
   ChatCompletionSchema,
