+++ backend/src/routes/private/api-keys.ts
@@ -3,8 +3,8 @@ import { FastifyInstance } from "fastify";
 import {
   ApiKeyCreateResponse,
   ApiKeysGetResponse,
-} from "gpinterface-shared/type/apiKey";
-import { createApiKey } from "../../controllers/apiKey";
+} from "gpinterface-shared/type/api-key";
+import { createApiKey } from "../../controllers/api-key";
 import { DeleteResponse, HashIdParam } from "gpinterface-shared/type";
 
 export default async function (fastify: FastifyInstance) {
