+++ shared/type/chat-content.ts
